pyproject.toml
[tool.poetry]
name = "link-shortener"
version = "0.1.0"
description = "API based URL shortener to generate and redirect links"
authors = ["Solomon Defi <solomondefi@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^1.1.2"
Flask-SQLAlchemy = "^2.5.1"
SQLAlchemy = "^1.4.6"
uWSGI = "^2.0.19"
python-dotenv = "^0.17.0"
ansicolors = "^1.1.8"

[tool.poetry.dev-dependencies]
alembic = "^1.5.8"
black = "^20.8b1"
pytest = "^6.2.3"
Flask-Migrate = "^2.7.0"
pylint = "^2.7.4"
pylint-flask = "^0.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
